+++ pages/demos/buildings.js
@@ -1,5 +1,5 @@
 import Head from 'next/head'
-import BuildingsDemo from '../../demos/buildings'
+import BuildingsDemo from 'demos/buildings'
 function Blur() {
   return (
     <>
