+++ pages/demos/tree-drawing.js
@@ -2,11 +2,11 @@ import {
   Algorithm1Demo,
   Algorithm2Demo,
   Algorithm3Demo
-} from '../../demos/tree-drawing/tidy-drawings-of-trees'
-import { TRAlgorithmDemo } from '../../demos/tree-drawing/tidier-drawings-of-trees'
-import { WalkerAlgorithmDemo } from '../../demos/tree-drawing/a-node-positioning-algorithm-for-general-trees'
-import '../../demos/tree-drawing/greatest-distinct-ancestors'
-import { ImprovedWalkerAlgorithmDemo } from '../../demos/tree-drawing/improving-walker-algorithm-to-run-in-linear-time'
+} from 'demos/tree-drawing/tidy-drawings-of-trees'
+import { TRAlgorithmDemo } from 'demos/tree-drawing/tidier-drawings-of-trees'
+import { WalkerAlgorithmDemo } from 'demos/tree-drawing/a-node-positioning-algorithm-for-general-trees'
+import 'demos/tree-drawing/greatest-distinct-ancestors'
+import { ImprovedWalkerAlgorithmDemo } from 'demos/tree-drawing/improving-walker-algorithm-to-run-in-linear-time'
 function TreeDrawing() {
   return (
     <>
