+++ pages/post/a+-promise.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '手写一个符合 A+ 规范的 Promise',
