+++ pages/post/avoid-cannot-read-property-of-undefined.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '避免取值时出现 Cannot read property \'xx\' of undefined',
