+++ pages/post/cesium-buildings.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: 'Cesium 实现建筑夜景贴图',
