+++ pages/post/debounce-and-throttle.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '探究防抖 (debounce) 和节流 (throttle)',
