+++ pages/post/doubly-linked-list-in-js.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: 'JavaScript 中的双向链表',
