+++ pages/post/fast-blur.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
 	title: '实现 O(n) 时间的高斯模糊算法',
