+++ pages/post/miniprogram-computed-watch.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '实现微信小程序的直接赋值、computed 和 watch 功能',
