+++ pages/post/npm-link.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '使用 npm link 遇到的一些坑',
