+++ pages/post/react-antd-with-ie8.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '记录 react+antd 兼容 ie8 的一些坑',
