+++ pages/post/read-the-shortest-history-of-europe.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '极简欧洲史读后',
