+++ pages/post/regex-vis.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '使用 regex-vis.com 可视化正则表达式',
