+++ pages/post/regex-vis_en.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: 'Use regex-vis.com to visualize regular expressions',
