+++ pages/post/rust-mdx.mdx
@@ -1,12 +1,33 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '用 Rust 重写 MDX 编译',
   time: '20220116',
   tags: ['Rust']
 }
+import {
+  BarChart,
+  CartesianGrid,
+  XAxis,
+  YAxis,
+  Tooltip,
+  Legend,
+  Bar,
+  ResponsiveContainer
+} from 'recharts'
 
-import MdxorChart from 'components/charts/mdxor.js'
+export const data = [
+  {
+    name: 'Official Compiler',
+    node: 225,
+    browser: 124
+  },
+  {
+    name: 'Mdxor Compiler',
+    node: 15963,
+    browser: 8508
+  }
+]
 
 [@mdxor/compiler](https://github.com/mdxor/compiler) 是一个用 Rust 编写、高速并且零依赖的 MDX 解析器。
 我刚刚完成它的绝大部分功能，这里分享一下它的实际效果和实现原理
@@ -26,7 +47,28 @@ Mdxor Wasm           x 8,181 ops/sec ±7.06% (46 runs sampled)
 Official Mdx Browser x 125 ops/sec ±23.35% (44 runs sampled)
 ```
 
-<MdxorChart />
+<div style={{ width: '100%', height: '300px' }}>
+	<ResponsiveContainer>
+		<BarChart
+			data={data}
+			layout="vertical"
+			margin={{
+				top: 5,
+				right: 30,
+				left: 20,
+				bottom: 5
+			}}
+		>
+			<CartesianGrid strokeDasharray="3 3" />
+			<XAxis type="number" />
+			<YAxis type="category" dataKey="name" />
+			<Tooltip />
+			<Bar dataKey="node" fill="#8884d8" />
+			<Bar dataKey="browser" fill="#82ca9d" />
+			<Legend />
+		</BarChart>
+	</ResponsiveContainer>
+</div>
 
 不论是从文字表述还是可视化图表，都可以看出在该项测试中 `@mdxor/compiler` 的速度大幅领先官方的 `@mdx-js/mdx`。
 其中，Node.js 上的速度约为官方的 70 倍，浏览器上则约为 65 倍
