+++ pages/post/simple-vue.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '实现一个简单的 Vue',
