+++ pages/post/simply-file-miniprogram.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '实现小程序的单文件开发模式',
