+++ pages/post/spell-check.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '实现拼写检查器 (spell check)',
