+++ pages/post/use-react-hooks-implement-redux.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '使用 react hooks 实现自己的 context-redux',
