+++ pages/post/vue-to-miniprogram-2.mdx
@@ -1,4 +1,4 @@
-import Layout from 'components/layout/post'
+import Layout from './_post'
 
 export const meta = {
   title: '如何实现 Vue 转小程序 (2)',
